+++ packages/decap-cms-core/src/components/App/Header.js
@@ -71,7 +71,7 @@ const AppHeaderButton = styled.button`
 
   &:hover,
   &:active,
-  &:focus {
+  &:focus-visible {
     ${styles.buttonActive};
 
     ${Icon} {
@@ -131,6 +131,9 @@ class Header extends React.Component {
   intervalId;
 
   componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(Header.propTypes, this.props, 'prop', 'Header');
+
     this.intervalId = setInterval(() => {
       this.props.checkBackendStatus();
     }, 5 * 60 * 1000);