+++ packages/decap-cms-core/src/components/Editor/EditorControlPane/EditorControl.js
@@ -5,7 +5,8 @@ import ImmutablePropTypes from 'react-immutable-proptypes';
 import { translate } from 'react-polyglot';
 import { ClassNames, Global, css as coreCss } from '@emotion/react';
 import styled from '@emotion/styled';
-import { partial, uniqueId } from 'lodash';
+import partial from 'lodash/partial';
+import uniqueId from 'lodash/uniqueId';
 import { connect } from 'react-redux';
 import { FieldLabel, colors, transitions, lengths, borders } from 'decap-cms-ui-default';
 import ReactMarkdown from 'react-markdown';
@@ -62,7 +63,6 @@ const styleStrings = {
   disabled: `
     pointer-events: none;
     opacity: 0.5;
-    background: #ccc;
   `,
   hidden: `
     visibility: hidden;
@@ -77,21 +77,26 @@ const ControlContainer = styled.div`
   }
 `;
 
+const ControlTopbar = styled.div`
+  display: flex;
+  justify-content: space-between;
+  gap: 20px;
+  align-items: end;
+`;
 const ControlErrorsList = styled.ul`
   list-style-type: none;
   font-size: 12px;
   color: ${colors.errorText};
-  margin-bottom: 5px;
   text-align: right;
   text-transform: uppercase;
-  position: relative;
   font-weight: 600;
-  top: 20px;
+  margin: 0;
+  padding: 2px 0 3px;
 `;
 
 export const ControlHint = styled.p`
   margin-bottom: 0;
-  padding: 3px 0;
+  padding: 6px 0 0;
   font-size: 12px;
   color: ${props =>
     props.error ? colors.errorText : props.active ? colors.active : colors.controlLabel};
@@ -135,7 +140,6 @@ class EditorControl extends React.Component {
     removeInsertedMedia: PropTypes.func.isRequired,
     persistMedia: PropTypes.func.isRequired,
     onValidate: PropTypes.func,
-    processControlRef: PropTypes.func,
     controlRef: PropTypes.func,
     query: PropTypes.func.isRequired,
     queryHits: PropTypes.object,
@@ -167,6 +171,11 @@ class EditorControl extends React.Component {
 
   uniqueFieldId = uniqueId(`${this.props.field.get('name')}-field-`);
 
+  componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(EditorControl.propTypes, this.props, 'prop', 'EditorControl');
+  }
+
   isAncestorOfFieldError = () => {
     const { fieldsErrors } = this.props;
 
@@ -197,7 +206,6 @@ class EditorControl extends React.Component {
       removeInsertedMedia,
       persistMedia,
       onValidate,
-      processControlRef,
       controlRef,
       query,
       queryHits,
@@ -241,28 +249,30 @@ class EditorControl extends React.Component {
               ${isHidden && styleStrings.hidden};
             `}
           >
-            {widget.globalStyles && <Global styles={coreCss`${widget.globalStyles}`} />}
-            {errors && (
-              <ControlErrorsList>
-                {errors.map(
-                  error =>
-                    error.message &&
-                    typeof error.message === 'string' && (
-                      <li key={error.message.trim().replace(/[^a-z0-9]+/gi, '-')}>
-                        {error.message}
-                      </li>
-                    ),
-                )}
-              </ControlErrorsList>
-            )}
-            <LabelComponent
-              field={field}
-              isActive={isSelected || this.state.styleActive}
-              hasErrors={hasErrors}
-              uniqueFieldId={this.uniqueFieldId}
-              isFieldOptional={isFieldOptional}
-              t={t}
-            />
+            <ControlTopbar>
+              {widget.globalStyles && <Global styles={coreCss`${widget.globalStyles}`} />}
+              <LabelComponent
+                field={field}
+                isActive={isSelected || this.state.styleActive}
+                hasErrors={hasErrors}
+                uniqueFieldId={this.uniqueFieldId}
+                isFieldOptional={isFieldOptional}
+                t={t}
+              />
+              {errors && (
+                <ControlErrorsList>
+                  {errors.map(
+                    error =>
+                      error.message &&
+                      typeof error.message === 'string' && (
+                        <li key={error.message.trim().replace(/[^a-z0-9]+/gi, '-')}>
+                          {error.message}
+                        </li>
+                      ),
+                  )}
+                </ControlErrorsList>
+              )}
+            </ControlTopbar>
             <Widget
               classNameWrapper={cx(
                 css`
@@ -305,7 +315,10 @@ class EditorControl extends React.Component {
               value={value}
               mediaPaths={mediaPaths}
               metadata={metadata}
-              onChange={(newValue, newMetadata) => onChange(field, newValue, newMetadata)}
+              onChange={(newValue, newMetadata) => {
+                onChange(field, newValue, newMetadata);
+                clearFieldErrors(this.uniqueFieldId); // Видаляємо помилки лише для цього поля
+              }}
               onValidate={onValidate && partial(onValidate, this.uniqueFieldId)}
               onOpenMediaLibrary={openMediaLibrary}
               onClearMediaControl={clearMediaControl}
@@ -320,7 +333,6 @@ class EditorControl extends React.Component {
               resolveWidget={resolveWidget}
               widget={widget}
               getEditorComponents={getEditorComponents}
-              ref={processControlRef && partial(processControlRef, field)}
               controlRef={controlRef}
               editorControl={ConnectedEditorControl}
               query={query}