+++ packages/decap-cms-widget-markdown/src/serializers/remarkSlate.js
@@ -1,4 +1,9 @@
-import { isEmpty, isArray, flatMap, map, flatten, isEqual } from 'lodash';
+import isEmpty from 'lodash/isEmpty';
+import isArray from 'lodash/isArray';
+import flatMap from 'lodash/flatMap';
+import map from 'lodash/map';
+import flatten from 'lodash/flatten';
+import isEqual from 'lodash/isEqual';
 
 /**
  * Map of MDAST node types to Slate node types.
@@ -234,12 +239,20 @@ export default function remarkToSlate({ voidCodeBlock } = {}) {
        * Convert simple cases that only require a type and children, with no
        * additional properties.
        */
-      case 'root':
       case 'paragraph':
       case 'blockquote':
       case 'tableRow':
       case 'tableCell': {
         return createBlock(typeMap[node.type], nodes);
+      }
+
+      /**
+       * Root element
+       * If the root node is empty, we need to add a paragraph node to it.
+       */
+      case 'root': {
+        const children = isEmpty(nodes) ? [createBlock('paragraph')] : nodes;
+        return createBlock(typeMap[node.type], children);
       }
 
       /**