+++ packages/decap-cms-widget-select/src/SelectControl.js
@@ -2,13 +2,15 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import ImmutablePropTypes from 'react-immutable-proptypes';
 import { Map, List, fromJS } from 'immutable';
-import { find } from 'lodash';
+import find from 'lodash/find';
 import Select from 'react-select';
 import { reactSelectStyles } from 'decap-cms-ui-default';
 import { validations } from 'decap-cms-lib-widgets';
 
 function optionToString(option) {
-  return option && option.value ? option.value : null;
+  return option && (typeof option.value === 'number' || typeof option.value === 'string')
+    ? option.value
+    : null;
 }
 
 function convertToOption(raw) {
@@ -47,9 +49,10 @@ export default class SelectControl extends React.Component {
       options: ImmutablePropTypes.listOf(
         PropTypes.oneOfType([
           PropTypes.string,
+          PropTypes.number,
           ImmutablePropTypes.contains({
             label: PropTypes.string.isRequired,
-            value: PropTypes.string.isRequired,
+            value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
           }),
         ]),
       ).isRequired,
@@ -94,6 +97,9 @@ export default class SelectControl extends React.Component {
   };
 
   componentDidMount() {
+    // Manually validate PropTypes - React 19 breaking change
+    PropTypes.checkPropTypes(SelectControl.propTypes, this.props, 'prop', 'SelectControl');
+
     const { field, onChange, value } = this.props;
     if (field.get('required') && field.get('multiple')) {
       if (value && !List.isList(value)) {